// src/app/model/MakeFeeCalc.model.ts
export interface MakeFeeCalcModel {
  day: string;
  amountOfInsurance: string;
  year: string;
  month: string;
  lengthOfInsurance: string;
  numberOfInsured: string;
  chargeFrequency: string;
  policyDiscount: string;
  paymentMethodDiscount: string;
  customerDiscount: string;
  campaignDiscount: string;
}